Loads-a-changes
Switches to yaml file for track meta data, which allows modeling of optional data and makes the content more legible.

Creates a Target class to describe a point on a track which we need to determine when the point has been crossed.

Refactors track.py and lap_tracker.py to take advantage of the above two changes.

Adds more panels to the GUI to display driver related information as the car enters the pits, and while the car is being serviced during a pit stop.

Refactors some more spatial calculations into geometry.py.