Lots more little fixes:
Reduces the display size a bit to accommodate the fact that we don't get the whole screen size.
Suppressed event logs now show the name of the event.
Fixes GUI so it registers itself for the events it needs.
Fixes angular difference calculation, and adds some tests.
Caches fuel used in the last hour, updating once a minute.