Disable wifi (#14)
* Big Fuel Usage Refactor

Moved the fuel consumption calculation into obd_reader.py and now it updates fuel listeners with the time and the rate of fuel being consumed.

The fuel rate is now more accurate, hopefully! As it uses fuel_trim rates in calculating the fuel flow.

moves tests into tests/ directory

* disables wifi in order to reduce battery draw.
that probably needs to only happen in prod mode
so coming soon will be some dev/prod settings
also, if the local time is wrong compared to gps, then we set the clock based on the gps provided time.