Handle messages from pit in car wires up the lemon-pi in car application to over the radio messages.
The GUI is tightened up vertically with some less useful bits removed.

A lower line is added to display general messages as well as position information

Also does some housekeeping of the event class in preparation for being able to move it into shared, and then just keep event definitions in the different applications.