adds support for different classes within the field (#138)
keeps track of position in class as well as overall
switches to use float times (in seconds) throughout the pit rather than string times which were coming from the datasource.
adds a suggested lap time to try to maintain ... v simplistic. Might help at the start of lemons. The suggested lap time is displayed in the pit gui (it is not sent to the driver)