protobuf cleanup (#139)
* protobuf cleanup
 removed proprietary implementation of 'oneof' by using the builtin mechanism instead.
 we now only send one of two message types over the radio :
    ToCarMessage, or
    ToPitMessage

This does mean that when we initialize a Radio instance we need to give it the base class
that it will use to decode all incoming messages. It also means that the wrong kind of message
cannot be received by either device (pit or car).

Overall this is a nice cleanup, as it's safer to add new message types in the future than the
old funky implementation, that required new message types all over the place.